requirements.txt
click==8.0.4
cloudpickle==2.1.0
cycler==0.11.0
dataclasses==0.8
Flask==2.0.3
fonttools==4.27.1
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.0
joblib==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mysql-connector-python==8.0.30
numba==0.53.1
numpy==1.19.5
packaging==21.3
pandas==1.0.5
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.4
pyparsing==3.0.9
python-dateutil==2.7.0
pytz==2022.1
scikit-learn==0.24.2
scipy==1.5.4
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
